implements PR feedback changes
